requirements.txt
tkinterdnd2>=0.3.0
gpxpy>=1.6.2
fitparse>=1.2.0
Pillow>=10.0.0
piexif>=1.1.3
pyinstaller>=6.0.0
